# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CareerRoadmapInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  degree: string;
  skills: string[];
  projects: string;
  /** @minLength 1 */
  career: string;
  customCareer: string;
  /** @minLength 1 */
  experience: string;
  /** @minLength 1 */
  learningTime: string;
}

export interface CareerProject {
  title: string;
  description: string;
  skills: string[];
}

export interface CareerRoadmapStep {
  week: string;
  focus: string;
  outcome: string;
}

export interface CareerRoadmap {
  career: string;
  requirements: string[];
  strengths: string[];
  gaps: string[];
  priorities: string[];
  /**
     * @minimum 0
     * @maximum 100
     */
  readiness: number;
  projects: CareerProject[];
  roadmap: CareerRoadmapStep[];
  recommendation: string;
}

export interface ErrorResponse {
  error: string;
}