# lib/api-zod/src/generated/types/careerProject.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerProject {
  title: string;
  description: string;
  skills: string[];
}