# lib/api-zod/src/generated/types/careerRoadmap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareerProject } from './careerProject';
import type { CareerRoadmapStep } from './careerRoadmapStep';

export interface CareerRoadmap {
  career: string;
  requirements: string[];
  strengths: string[];
  gaps: string[];
  priorities: string[];
  /**
     * @minimum 0
     * @maximum 100
     */
  readiness: number;
  projects: CareerProject[];
  roadmap: CareerRoadmapStep[];
  recommendation: string;
}