# lib/api-zod/src/generated/types/careerRoadmapInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerRoadmapInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  degree: string;
  skills: string[];
  projects: string;
  /** @minLength 1 */
  career: string;
  customCareer: string;
  /** @minLength 1 */
  experience: string;
  /** @minLength 1 */
  learningTime: string;
}