# lib/api-zod/src/generated/types/careerRoadmapStep.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerRoadmapStep {
  week: string;
  focus: string;
  outcome: string;
}